requirements/local.txt
-r ./base.txt

Werkzeug==0.16.1 
ipdb==0.12.3  
Sphinx==2.3.1  
psycopg2==2.8.4
mypy==0.761  
django-stubs==1.4.0  
pytest==5.3.5  
pytest-sugar==0.9.2  
flake8==3.7.9  
coverage==5.0.3  
black==19.10b0  
pylint-django==2.0.13  
pylint-celery==0.3  
pre-commit==2.0.1  
factory-boy==2.12.0  
django-debug-toolbar==2.2  
django-extensions==2.2.6  
django-coverage-plugin==1.8.0  
pytest-django==3.8.0  
PyYAML==5.3.1